requirements.txt
PyQt5>=5.15.0
pyyaml>=6.0
pyinstaller>=5.13.0
